% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ch_init_schema}
\alias{ch_init_schema}
\title{Create ClickHouse tables if they do not exist}
\usage{
ch_init_schema(con)
}
\arguments{
\item{con}{A \code{DBIConnection} from \code{\link{ch_connect}}.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Create ClickHouse tables if they do not exist
}
