% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ch_topic_summary}
\alias{ch_topic_summary}
\title{Get aggregated topic statistics}
\usage{
ch_topic_summary(con)
}
\arguments{
\item{con}{DBI connection.}
}
\value{
Data frame with columns \code{topic_label}, \code{n_articles},
\code{first_article}, \code{last_article}.
}
\description{
Get aggregated topic statistics
}
