% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{ch_write_articles}
\alias{ch_write_articles}
\title{Write articles data frame to ClickHouse}
\usage{
ch_write_articles(con, df)
}
\arguments{
\item{con}{DBI connection.}

\item{df}{Data frame produced by \code{\link{fetch_news_dataframe}} and
optionally \code{\link{classify_news}}.}
}
\value{
\code{df} invisibly.
}
\description{
Write articles data frame to ClickHouse
}
