% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ttrss_get_api_level}
\alias{ttrss_get_api_level}
\title{Return the TT-RSS API level}
\usage{
ttrss_get_api_level(base_url, session_id)
}
\arguments{
\item{base_url}{Base URL.}

\item{session_id}{Session ID from \code{ttrss_login()}.}
}
\value{
Named list with \code{level}.
}
\description{
Return the TT-RSS API level
}
