% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ttrss_get_article}
\alias{ttrss_get_article}
\title{Fetch full article content for one or more article IDs}
\usage{
ttrss_get_article(base_url, session_id, article_ids)
}
\arguments{
\item{base_url}{Base URL.}

\item{session_id}{Session ID from \code{ttrss_login()}.}

\item{article_ids}{Integer vector of article IDs.}
}
\value{
A data frame with full article fields.
}
\description{
Fetch full article content for one or more article IDs
}
