% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ttrss_get_categories}
\alias{ttrss_get_categories}
\title{Get all feed categories}
\usage{
ttrss_get_categories(base_url, session_id)
}
\arguments{
\item{base_url}{Base URL.}

\item{session_id}{Session ID from \code{ttrss_login()}.}
}
\value{
A data frame of categories (id, title, unread, order_id).
}
\description{
Get all feed categories
}
