% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ttrss_get_feeds}
\alias{ttrss_get_feeds}
\title{Get feeds, optionally filtered by category}
\usage{
ttrss_get_feeds(base_url, session_id, cat_id = -1)
}
\arguments{
\item{base_url}{Base URL.}

\item{session_id}{Session ID from \code{ttrss_login()}.}

\item{cat_id}{Category ID (\code{-1} = all feeds, \code{-2} = uncategorised).}
}
\value{
A data frame of feeds.
}
\description{
Get feeds, optionally filtered by category
}
