% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ttrss_logout}
\alias{ttrss_logout}
\title{Log out of TT-RSS}
\usage{
ttrss_logout(base_url, session_id)
}
\arguments{
\item{base_url}{Base URL.}

\item{session_id}{Session ID from \code{ttrss_login()}.}
}
\value{
\code{TRUE} invisibly.
}
\description{
Log out of TT-RSS
}
